cc_light resource manager extended to support backward resource management, e.g. in combination with ALAP (backward compatible change)
The cc_light resource manager has been extended with backward resource management next to forward resource management.
This supports ALAP_rc scheduling next to ASAP_rc scheduling as until now.
Furthermore, the resource manager was extended with the new resource detuned_qubits.
And the resource manager was extended with several constructors and an improved destructor.
A new .cc test series was added and an accompanying .json file (with the new resource constraint).

This change to the resource manager is backward compatible.
The old use and old constructor creates a forward scheduling resource manager.

A bug in qwg resource management was solved.
However, it could not be reproduced until now.

NOTE: The ALAP rc scheduling was not added yet.