Print kernel statistic 'qubit cycles use', for each qubit the sum of the number of cycles it was busy executing gates
Each qubit cycles use value, when divided by Depth, gives a 0<=value<=1, giving fraction of cycles qubit was not idling.
When summed and divided by the number of used qubits, gives a measure of the parallelism of the circuit, 1==full.
This measure indicates how well the chain of circuit, mapper, and scheduler are performing.
Also when reading this value after each pass, and dividing the value after by before, it gives a measure for the pass.