ASAP/ALAP scheduling without rc done, with rc and uniform to be done; separate entry points for pre179 and post179
A new/temporary option scheduler_post179 with values no/yes, default no, allows to switch
between the old code and the new (after solving issue 179) code. So default is the old behavior.
Apart from changes to scheduler.h and options.h,
the gate definition was extended with a cycle field to store the resulting cycle in which a gate was scheduled.
This allows the bundler and other passes after scheduling to be done after deallocation of the scheduler.