Deep-critical path scheduling (ASAP and ALAP) working.
The commutativity support in the dependence graph unleashed a lot of freedom to the scheduler:
it has more options to choose from (all in its avlist) and more of them are critical (delaying one would delay the circuit).
So just comparing two options for the amount of criticality is not sufficient anymore:
also the criticality of the depending nodes and even the whole critical depending dependence subgraph behind each matters.
This criticality comparison for selecting the best option to schedule next has been extended to look deep forward.
This was done by keeping the avlist in deep-critical order, highest deep-critical node first.