Added mngmt of qubits (mapping and real state) and allow state to be passed on from start to initplace to heuristics
Past is not used after MapCircuit returns, it is purely local; ready to be moved further down.
  Replaced calls to retrieve statistics by intermediate step in mapper's state.
Added state with virtual qubit that it may not be mapped; added checks/asserts on it.
  Added option to have initial mapping one2one or undefined (mapinitone2one option).
Added state with real qubit that it is garbage/inited/hasstate; added checks/asserts on it.
  Adapted generation of moves, generation of very first initial one2one mapping and initial placement to this.
  Put real qubit state inside Virt2Real.
Turned Virt2Real from local to Past to a parameter of Initial Placement and heuristics,
  so that it can be passed between them, updating it on the fly, in preparation of multiple kernel mapping.
  But it also took the Virt2Real flow mgmt away from Past; Past just uses it, allowing Past to be cleaned further.
Improved printing of Virt2Real
Explained flow of a gate through: creation, swap gen, devirtualization, decomposing, scheduling wrt json file.
  In .json files, removed non-composite _real gates, added plain move/swap gates, matching these explanations.