Inconsistent use of resource constraints in past ASAP schedule resolved. May have led to wrong schedules.
past.Schedule used tryfc.AddNoRc instead of tryfc.Add, but used tryfc.StartCycle: inconsistent!
Disentangled use of NoRc versus withrc in past.scheduling.
Added option mapreverseswap.
Added check in fc to see which of two qubits can be scheduled to first,
to steer whether to do a swap or a reversed swap.
A sequence of swaps is assumed to all be reversed or not reversed.
This ignores resource constraints and ignores other gates executing in involved qubits.