Added report generation to quantumsim compiler and in generated python script.
Copied more or less the report generation from cc_light to quantumsim.
The report printed by the generated python script for quantumsim needed
initial spacing on each line so support was added for that to the report facility.

An addition option mapselectmaxswaps was added to maximize
the number of swaps inserted when selecting an alternative.
This means that the target CNOT may not get NN when this max is set to 1.
And that not all swaps are added in one go for one target CNOT,
but that for a set of CNOTs, distributed over these, gradually single swaps are added,
with the idea that they interfer and give a better result in total.
Note that this requires that maplookahead is set to "all".
Perhaps we need to extend the latter option with a value, e.g. to select the most critical 2.