Kernel.h diff with develop reduced.
The diff of kernel.h between mapper and develop branch was further reduced
by taking out an example new gate implementation from the mapper's kernel.h.
All kinds of other quirks were removed except for those proposed to be merged into develop.
In the mapper, the implementation of the option mapreverseswap was wrong; it was reimplemented.
Generating a move was factored out of AddSwap.
The interfaces of the reporting facility in report.h were updated preventing confusion with existing interfaces.