Standardized on space separating operands in json because of various inconsistencies.
Standardized on space separating operands in json because of various inconsistencies:
src/hardware_configuration was inconsistent wrt real instructions and constructed sub instructions.
Adapted src/kernel.h and src/mapper.h accordingly.
Wrong is the coupling of getting gate parameters (in src/scheduler.h) to what is in json and not to what is in instruction map;
i.e. the name of the gate is used to find the json entry; this name was the result of various updates (lower-case, space removal,
and now: comma removel).
Furthermore it is strange that a subins like blue q0 q1 cannot be matched by an instr like blue; a gate could be made for it;
but getting its gate parameters wouldn't work ...

Added a lot of DOUTs around this.

Mapper: added support for adding just the in-time first possible swap instead of first generated swap.
And resolved a bug in reverseswap.
Removed support for adding 2, 3, ... number of swaps; kept first, earliest, all; old one was "all".

Bug also in MakeReal/MakePrimitive: suffix _real/_prim was added to name, where name might be specialized;
so now it is added to gate name in name, i.e. omitting the operands and then the thing is tried.

Improved the error reporting on wrong topology specification in the json file.

Added test_rig.cc and test_rig.json to demonstrate rigetti support and try mapping a circuit for it.