Enabled gate creation thru kernel.gate_nonfatal instead of mapper.new_gate
Enabled gate creation thru kernel.gate_nonfatal instead of mapper.new_gate.
This makes copy inside mapper of kernel gate implementation unused; perhaps some comments need to be copied before deletion.
The mapper was updated to provide access to the kernel class where needed to create a gate.
For this, the input kernel.c was copied out of the way so that kernel.c is free for use by gate creation.
Gate creation in the mapper lets them be created in kernel.c by kernel.gate_nonfatal and
then copies them out to the mapper.new_gate's circ reference parameter.
It is taken care of that kernel.c is empty where needed; some asserts for this may still be deleted.

Furthermore, test_mapper was enabled again instead of test_rig.