Replaced size_t cycleExtend by double score as field of class Alter.
Replaced cycleExtend attribute of each Alter by score, of type double.
The type double also supports a score that is the fidelity of the circuit,
where size_t was used because cycleExtend was just a (positive) difference of cycle values (which have type size_t).
Use a new field (bool didscore) to know whether a significant value was stored in score, for printing (replacing MAX_CYCLE as value of cycleExtend).
In src/eqasm_compiler.h, update the kernels parameter of the virtual compile method to be by reference (a bug in the previous commit).