Deleted src/arch/quantumsim_compiler.h; deleted quantumsim related code from src/program.h
Deleted src/arch/quantumsim_compiler.h
All remains of it (#include and checking in src/program.h) were deleted.
Note that the code to write a quantumsim program for cc_light was already moved to the end of src/arch/cc_light/cc_light_eqasm_compiler.h.
When it turns out that that code is cc_light independent, it must be extracted to src/quantumsim.h