Added unique_output option to make output qasm/report files unique on creation
The new option has value "no" or "yes", default "no".
When set, each output qasm/report file with a name derived from the
program name is made unique by appending a version number to its name.
To this end, a file with name <test_output_dir>/<program_name>.unique
is maintained that contains the version number of the last run.
It is created when not there.