Quantum operands in gate.operands, classical operands in gate.creg_operands, etc.
Classical operands were for most of the gates stored in gate.operands.
This led to problems in correct dependence analysis in the scheduler
and in correctly creating qasm for gates.
Before this change, several tests surprisingly didn't pass.
But after this pass, also several tests didn't pass anymore.
So also some of these passes were corrected,
e.g. to cater for the use of r28 in computing condition code which
requires the number of classical registers to be at least 29.
Still several tests don't pass, and this has to be analyzed.
Most problematic are the classical_cc gates that are put in a circuit
and subsequently a scheduler or a mapper has to cope with:
e.g. imagine an fmr that is unknown to the scheduler and for which
the dependencies have to be determined in the depgraph computation of the scheduler.
Possibly, the golden outputs have to be updated.

So, this commit is not done because the branch is consistent and fault-less.

Furthermore, as most user visible change, the interface to quantumsim has changed.
Quantumsim is not enabled by passing a quantum.json file as config file.
Instead, one uses any json file that is for a target that one wants to simulate,
and sets the "quantumsim" option to "no", "yes", or "qsoverlay";
the "yes" enables the usual quantumsim output, the "qsoverlay" for the qsoverlay version of quantumsim.
And furthermore, two outputs are produced, one after the prescheduler
and one just before writing the qisa file in cc_light.