In OpenQL's ir, kernel.c is the only construct that contains gates; kernel.bundles has been deleted.
Kernel.c and kernel.bundles were alternative structures to represent the circuit.
Now, kernel.c is the only one.
Where necessary locally in passes, bundler() creates bundles from a circuit and circuiter() creates circuit from bundles.
Whether bundles can be created, depends on kernel.cycles_valid.
This bool state value is set to false everywhere where a gate is added to or deleted from or moved in a circuit.
It is set to true everywhere where the circuit has been scheduled (so that the cycle attribute got valid,
and the circuit's gates have been sorted in non-decreating order of their cycle attribute).

Where necessary, code was moved/reorganized to make this possible.
The resulting implementation is fully backward compatible.

Some renaming was done to variables to make clear that they don't contain bundles.