Uniform scheduler is run when just scheduler_uniform=="yes", i.e. ignoring value of scheduler option
Whether uniform scheduling is ASAP or ALAP is a matter of taste. It is neither.
Also, we implemented only a single choice, i.e. one that has a uniform result
with a bias towards being ALAP.
Since we don't foresee a uniform scheduler with ASAP bias,
ignoring the scheduler option when uniform scheduling is easier towards the user.