Updated tests after remove CBox & add modularity
The configuration file for some of the tests needed to be updated
after the removal of the CBox backend. Furthermore, with the addition
of modularity some of the tests were failing in CI due to wrong options
being propagated from one unit test to another. Therefore, a reset_options
function has been added to the compiler that require now each unit test to call
explicitly the selfSetUp function to set its required options.

Changes to be committed:
	modified:   src/options.h
	modified:   tests/test_Platform.py
	modified:   tests/test_Program.py
	modified:   tests/test_alap_rc_schedule.py
	modified:   tests/test_barrier.py
	modified:   tests/test_bugs.py
	modified:   tests/test_cc_light.py
	modified:   tests/test_configuration.py
	modified:   tests/test_custom_gate.py
	modified:   tests/test_hybrid.py
	modified:   tests/test_minimal.py
	modified:   tests/test_modularity.py
	modified:   tests/test_options.py
	modified:   tests/test_qubits.py
	modified:   tests/test_sweep_points.py
	modified:   tests/test_unitary.py