Put initial place behind feature flag, default off
Initial placement requires lemon Mip, which (effectively) requires GLPK,
which is GPL, and this makes everything GPL by its copyleft principles.
This is probably not desirable in general, so the flag is off by
default, and in distributed packages it will be off for now. There's a
way to enable support for both CMake directly and for Python setup.
Checks for GLPK are conditioned on whether initial placement is enabled.