Fix setup.py
The license field is only to be used when there is no trove classifier
for the license (there is, and it's already in there), and if used,
must be a single line. The wheel metadata is horribly broken at the
moment, as evidenced by the project page at
https://pypi.org/project/qutechopenql/
(note the description and platform tags, and the actual description
being interpreted as a code block. this is not what it should be
looking like)