Added test_multi_core.py and testing qasm after code generation
Added test_multi_core.py and created a golden qasm for it.
However, the qasms generated by default don't include a qasm after code generation.
So this was added, under the name 'programname'_last.qasm.
See test_multi_core.py at its end.

A writer pass was added to the cc_light back-end to write this qasm.
The report::write_qasm facility was updated to recognize this writing;
this is a hack until options are available for that.

There was a bug in src/options.h with a wrong name for the option mapassumezeroinitstate.
After this correction, the python test assumes that all qubits are initialized in |0> state
so that moves/tmoves are generated where ever possible.

The C++ test tests/test_multi_core.cc was updated on the fly
but it is obsolete now, because the .py test does the same.
It is kept for future updates.