Added commute_variation pass
Ported and merged in the old code of the smallqec branch so that branch can be deleted.
That code was the result of varying the order of CNOTs etc in circuit exploiting their commutativity
especially in short error correction code circuits done by Lingling.

While the scheduler with its dependence graph exploits commutativity of CNOTs and CZs locally,
this pass exploits it globally in the whole circuit, i.e. choices of ordering the CNOTs etc.
of a single set of such gates, are made dependent on choices of ordering in other sets of such in the circuit.
After this, the exploitation of commutativity in the scheduler is turned off.
This pass serves as a prepass to the ultimate (resource-constrained) scheduler.