Resolve conflict between Grid struct/classes
Apparently the linker doesn't seem to care, but construction is
completely wrecked by two struct/classes with the same name in different
compile units that just never see each other's headers. This is why
namespaces exist. I only namespaced the mapper for now because it was
easy to do that, but in my later refactoring namespaces will be used
more extensively.