Identified bug in commute_variation regarding interplay between D and R dependences; added tests for those; not solved.
The following program exhibits an RAD dependence on q2:
    CNOT q[0],q[2]
    CNOT q[2],q[5]
This dependence implies non-commutability but this is not seen by commute_variation.
See test_cnot_noncommute_RAD.

Similarly, the following is also suspect then exhibiting a DAR dependence on q2:
    CNOT q[2],q[5]
    CNOT q[0],q[2]
See test_cnot_noncommute_DAR.

Test test_cnot_variations exhibits this in a larger context and so fails as well.