Disabled commute_variation pass and updated commute tests
Commute_variation pass is depending on depgraph implementation
and the latter is in flux now, so when it is stable again
this pass can be enabled again after an update.
Furthermore, tests for commutation entered the back-end but that was not needed for the test;
the tests have been simplified.