Updated tests/golden to avoid smart mapper/clifford effects to make those tests trivial.
Several tests started to fail because the mapper got to smart (it doesn't swap two qubits without state).
To prevent this, several of those and similar tests (in anticipation on more smartness) were updated.
Also, the excess setting of write_qasm_files to yes was removed; this option is a debugging aid.