Add support for limited number of communication qubits per core (topology.comm_qubits_per_core).
Inter-core tswap/tmoves are only possible through communication qubits at either side.
In the implementation, only these qubits are coupled/neighbours.
By default, all qubits in a core are communication qubits.
The number, when specified in the configuration file, must be positive and less-equal the number of qubits per core.
When specified, the first such qubits of each core become communication qubits; the remainder then are not.