Added channels resource to constrain number of parallel inter-core gates per core; default no constraint.
In config file, in resources section, added "channels" section to describe channels.
The "count" indicates maximum number of channels per core that may be busy with an inter-core gate.
Default (when resource not specified or count not specified) is that sufficient channels are available.
Please note that for each inter-core gate, a channel must be allocated in the core of each gate operand.
This resource is an alternative to the number of communication qubits per core configured in the topology section.
Indeed either setting to a particular value n should have the same effect; setting to nqubits/ncores behaves as no-op.