Added support for "extern" operation type (next to "mw","flux", etc.) to represent inter-core and similar "extern" operations
It allows to define additional gates next to tswap/tmove to be inter-core operations in the configuration file.
For those gates, specify the "type" to be "extern".
This means, those gates compete for communication channels in the resource constrained scheduling
and not for access to internal wave- or flux- or readout-interfaces.
This is not accurate for composite gates (or gates to be decomposed later); then take the most representative type that applies.

In the configuration file tests/test_multi_core_4x4_full.json this was updated.
tswap and tmove are still inserted by the mapper when detecting an inter-core hop;
but when generated, these are decomposed to more primitive gates such as preswap, teleportswap and postswap.
Those preswap/postswap/premove/postmove are defined now as single qubit gates;
when they would be multi qubit gates, then they should as well decompose to primitive gates;
when among those is a two qubit gate, its operands must be neighbours in the coupling graph;
when for this, an additional qubit would be needed, this would imply local qubit allocation
and that is not supported, so is a problem right now.
The teleportswap/teleportmove are meant as the real inter-core gates, so now are "extern" typed.

Note the sequence in the mapper thus is:
loop over gates until all done:
- it encounters a two-qubit gate
- when neighbours, continue
- find a path between them, and generate for all transitions except one:
  - move, when intra-core and target qubit has state |0>
  - swap, when intra-core and target qubit doesn't have state |0>
  - tmove, when inter-core and target qubit has state |0>
  - tswap, when inter-core and target qubit doesn't have state |0>
  and put the mapped two-qubit at the remaining transition
- when generating a move, swap, tmove, tswap:
  look it up in the config file, it can be a primitive or a composite;
  when move is not there, also try move_real, etc.
  when it is a composite, immediately decomposse it to more primitive gates;
  please note that this decomposition is not recursive,
  and that all more primitive gates when multi-qubits must be mappable (so neighbours, etc.)
  and that no additional qubits can be allocated during this decomposition (not supported (yet))
This produces a new circuit that conforms to coupling graph.

It is given to the resource-constrained scheduler (rcscheduler):
- a dependence graph is created between them, with dependences attributed with source gate duration
- cycles are filled with 'ready' gates in a greedy way
- a gate only fits in a cycle when its predecessors have completed and its resources are available;
  here for each gate the gate type is used to know which resources are needed;
  for those pre/post-swap/move those are "mw" (microwave rotation), "flux" (two-qubit gates), or "readout" (measurement)
  and for the real teleport those are "extern".
- when no gates remain to be filled in a cycle, start filling the next cycle;
  all gates filled in the cycle are set to 'scheduled' and the new 'ready' set
  are exactly those not yet scheduled but with only scheduled predecessors in the depgraph
- when all gates have been 'scheduled', ready.