Update mapper golden files (message for details)
As per #406, the mapper previously abused the resources in a way that
they're not designed for. The *new* resource implementation is, so
that'll obviously change test output. This commit adjusts resource
behavior such that out-of-order cycle availability checks always
return unavailable, so we should have consistent behavior when we
switch to the new implementation in a later commit. Once that's done,
the mapper can eventually be updated to use the UNDEFINED direction,
at which point things start making sense again, and may even be more
optimal than before.