Scheduler::criticality_lessthan simplified by taking recursion out LOG_IF_DEBUGs mostly disabled
Scheduler::criticality_lessthan simplified by taking recursion out:
when the criticality of two nodes is equal, the size of the lists of depending nodes decides: the longer, the more critical
otherwise, the original order is used (when earlier than more critical)
when global option scheduler_heuristic == "path_length" the above heuristic is taken
when global option scheduler_heuristic == "random" only the criticality value is taken, not the size of lists of dep nodes
Please note: ListScheduler was not updated with this (wait for Scheduler to become stable).

LOG_IF_DEBUGs mostly disabled:
That output poluted the text too much and overview was lost; please enable again when needed.