Added support for limiting the number of channel resources system-wide.
In the old resource specification, one can add "system_wide_count"
        "channels":
        {
            "description": "...",
            "count": 4,
            "system_wide_count", 2
        }
without it, for compatibility, the platform's number of qubits is taken as value for it.

In the new resource specification, one can add "num_system_wide_channels"
        {
            "predicate": { "type": "extern" },
            "inter_core_required": true,
            "communication_qubit_only": false,
            "num_channels": 4,
            "num_system_wide_channels": 2
        }
without it, for compatibility, the platform's number of qubits is taken as value for it.

Each gate matching the conditions (such as type is extern) needs
one channel for each qubit operand.
With a tswap, there are thus 2 channels associated, one in each core.
And tswap adds those 2 to the system-wide number of channels.