Limit check system-wide channels improved, channel saturation test added, goldens updated
Limit check system-wide channels updated to be independent of number of qubit operands.
Previously it implicitly assumed 2 operand qubits, now it can be also 1 or 3.

Updated test_multi_core_4x4_full.json with a new channel saturation test.
And with limiting bounds on channels per core (2) and on channels system-wide (4)
Consequently, the golden test output needed to be updated and verified.

Lots of debugging improvements