Replaced QL_IF_LOG_DEBUG idiom use by MULTI_LINE_LOG_DEBUG idiom to control multi-line debugging output
New idiom:

// #define MULTI_LINE_LOG_DEBUG to enable multi-line dumping

...

    QL_IF_LOG_DEBUG {
        QL_DOUT("Result of old->new IR platform conversion:");
        ir->dump_seq();
    }
    QL_DOUT("Result of old->new IR platform conversion (disabled)");

In the above, 'ir->dump_seq()' is not guarded internally by LOG_DEBUG/QL_DOUT
and prints the full IR to the debugging log.
To signal that this printing is prepared for but is currently disabled,
the message after the #else is given.

What we really want is a family of options to turn this on/off per file.
Per module (??) or per class is too cumbersome.