Trying fo fix cpp-linux-arm64 job.
It seems that cleaning the Conan cache is effective.
I have added it as a preventive measure to the cpp-tests/actions.yaml.
This will only be of some effect for the self-hosted runners.
But, since the self-hosted runners use containers that cache Conan packages from previous runs,
and that can cause different type of problems,
it looks like a sensible setup step to perform.